Seprate reusable newPasswordForm component
